backend/requirements.txt
# OkCupid Web API client (reverse-engineered)
requests[socks]>=2.28.0
cloudscraper>=1.2.71
python-dotenv>=1.0.0
openai>=1.60.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.4.0
